Refactored backend to use Bitfinex websockets
IEX trade data was not actually using WebSockets and doesn't provide updates when markets are closed, somewhat defeating the purpose of this demo.

Therefore I changed the data source to monitor cryptocurrency trading via the Bitfinex API and refactored the backend to use native WebSockets instead of Socket.io.

Also finished the basic functionality of the ticker component and fleshed out the app layout and controller, adding Add and Remove functionality and some other useful indicators to highlight trading activity.